requirements.txt
flask
rembg
pillow
gunicorn
flask-cors
onnxruntime